FLINK-2740 Fixing formatting issue with spaces, renaming topolgy classes to include Example in their name, and adding a license to the template file to pass rat check
