+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-connector-nifi/src/main/java/org/apache/flink/streaming/connectors/nifi/NiFiSinkTopologyExample.java
@@ -16,39 +16,35 @@
 */
 package org.apache.flink.streaming.connectors.nifi;
 
-import org.apache.flink.api.common.functions.MapFunction;
-import org.apache.flink.streaming.api.datastream.DataStream;
+import org.apache.flink.api.common.functions.RuntimeContext;
+import org.apache.flink.streaming.api.datastream.DataStreamSink;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.streaming.api.functions.source.SourceFunction;
 import org.apache.nifi.remote.client.SiteToSiteClient;
 import org.apache.nifi.remote.client.SiteToSiteClientConfig;
 
-import java.nio.charset.Charset;
+import java.util.HashMap;
 
 /**
- * An example topology that pulls data from a NiFi output port named "Data for Flink".
+ * An example topology that sends data to a NiFi input port named "Data from Flink".
  */
-public class NiFiSourceTopology {
+public class NiFiSinkTopologyExample {
 
 	public static void main(String[] args) throws Exception {
 		StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
 
 		SiteToSiteClientConfig clientConfig = new SiteToSiteClient.Builder()
 				.url("http://localhost:8080/nifi")
-				.portName("Data for Flink")
+				.portName("Data from Flink")
 				.buildConfig();
 
-		SourceFunction<NiFiDataPacket> nifiSource = new NiFiSource(clientConfig);
-		DataStream<NiFiDataPacket> streamSource = env.addSource(nifiSource); //.setParallelism(2);
+		DataStreamSink<String> dataStream = env.fromElements("one", "two", "three", "four", "five", "q")
+				.addSink(new NiFiSink<>(clientConfig, new NiFiDataPacketBuilder<String>() {
+					@Override
+					public NiFiDataPacket createNiFiDataPacket(String s, RuntimeContext ctx) {
+						return new StandardNiFiDataPacket(s.getBytes(), new HashMap<String,String>());
+					}
+				}));
 
-		DataStream<String> dataStream = streamSource.map(new MapFunction<NiFiDataPacket, String>() {
-			@Override
-			public String map(NiFiDataPacket value) throws Exception {
-				return new String(value.getContent(), Charset.defaultCharset());
-			}
-		});
-
-		dataStream.print();
 		env.execute();
 	}
